Move some components into their own subfolder
