+++ src/components/NewLoginModal.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import ActionButton from "./ActionButton.vue";
+import ActionButton from "./buttons/ActionButton.vue";
 import ConfirmGotNewAccountId from "./ConfirmGotNewAccountId.vue";
 import Modal from "./Modal.vue";
 import OutLink from "./OutLink.vue";
