+++ src/components/attachments/FileView.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import type { Attachment } from "../../model/Attachment";
 import type { PropType } from "vue";
-import ActionButton from "../ActionButton.vue";
-import DownloadButton from "./DownloadButton.vue";
+import ActionButton from "../buttons/ActionButton.vue";
+import DownloadButton from "../buttons/DownloadButton.vue";
 import ErrorNotice from "../ErrorNotice.vue";
 import List from "../List.vue";
 import TransactionListItem from "../transactions/TransactionListItem.vue";
