+++ src/components/settings/ChangePasswordForm.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import ActionButton from "../ActionButton.vue";
+import ActionButton from "../buttons/ActionButton.vue";
 import TextField from "../TextField.vue";
 import { computed, ref } from "vue";
 import { useAuthStore } from "../../store/authStore";
