+++ src/components/settings/ConfirmDeleteEverything.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import ActionButton from "../ActionButton.vue";
+import ActionButton from "../buttons/ActionButton.vue";
 import Confirm from "../Confirm.vue";
 
 const emit = defineEmits(["yes", "no"]);
