+++ src/components/settings/DeleteData.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import ActionButton from "../ActionButton.vue";
+import ActionButton from "../buttons/ActionButton.vue";
 import ConfirmDeleteEverything from "./ConfirmDeleteEverything.vue";
 import TextField from "../TextField.vue";
 import { computed, ref } from "vue";
