+++ src/components/settings/Import.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import type { DatabaseSchema } from "../../model/DatabaseSchema";
 import type { Entry } from "@zip.js/zip.js";
-import ActionButton from "../ActionButton.vue";
+import ActionButton from "../buttons/ActionButton.vue";
 import FileInput from "../attachments/FileInput.vue";
 import ImportProcessModal from "./ImportProcessModal.vue";
 import { BlobReader, TextWriter, ZipReader } from "@zip.js/zip.js";
