+++ src/components/tags/ConfirmDestroyTag.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import type { PropType } from "vue";
 import type { Tag } from "../../model/Tag";
-import ActionButton from "../ActionButton.vue";
+import ActionButton from "../buttons/ActionButton.vue";
 import Confirm from "../Confirm.vue";
 import { computed, toRefs } from "vue";
 import { useTransactionsStore } from "../../store";
