+++ src/router/index.ts
@@ -3,11 +3,11 @@ import About from "../About.vue";
 import Accounts from "../components/accounts/Accounts.vue";
 import AccountView from "../components/accounts/AccountView.vue";
 import Attachments from "../components/attachments/Attachments.vue";
-import EmptyRoute from "../components/EmptyRoute.vue";
+import EmptyRoute from "../pages/EmptyRoute.vue";
 import Home from "../Home.vue";
 import Install from "../Install.vue";
 import Locations from "../components/locations/Locations.vue";
-import Login from "../components/Login.vue";
+import Login from "../pages/Login.vue";
 import Security from "../Security.vue";
 import Settings from "../components/settings/Settings.vue";
 import Tags from "../components/tags/Tags.vue";
